AI/agents/requirements.txt
langchain-core
langchain-openai
langgraph
cdp-langchain
cdp
pydantic
twitter-langchain
langchain-community
duckduckgo-search
python-dotenv